refactor: merge delete and update refresh paths into session_product_refresh.py
Co-locate the delete-side and update-side maintenance logic so the incremental refresh contract is readable from one module.
